+++ CHANGELOG.md
@@ -8,7 +8,14 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 = 2.7.2 =
+* Updated: Freemius SDK (2.13.1)
 * Fixed: Early capability settings check (external plugin compat)
+* Fixed: Apply filter to check host/producer users still exist
+* Fixed: Override Specials data linked show meta
+* Fixed: Schedule Engine override ID key bug
+* Added: Override Specials content data for single/multiple
+* Added: Admin plugin Settings page styles
+* Added: Station Genre field (for single genre stations)
 
 = 2.7.1 =
 * Fixed: Print Late Footer Stylesheets Priority